You can speed up the code compile/link/debug cycle with Bakefile. This example focuses on compiling a GUI executable in Linux using Bakefile. The Bakefile program produces a GNUMakefile that has the instructions on compiling code written in C. When make is run, the executable is created.
The example executable that is created uses GTK 3 and is based entirely on the GTK 3 Hello World example in Part 1 of the GTK manual for 2019. I included a PDF snapshot of the GTK manual in the Reference Manual subdirectory.

I made custom versions of the manuals for Bakefile and GTK 3 using pandoc to odt conversion. I opened the generated odt files, cleaned them up, and exported them to PDF from within LibreOffice.

A full write-up on the process can be found on the blog: https://gautiertalkstechnology.wordpress.com on 9/23/2019 at 4AM CST.